add authorization and characteristics

+++ app/routes/categories.py
@@ -1,8 +1,9 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException
 
 from app.models.category import Category
 from app.models.product import Product
 from app.schemas.categories import CategoryCreateModel, CategoryUpdateModel
+from app.utils import Permissions, AuthManagerDep
 
 router = APIRouter(prefix="/api/v0/categories")
 
@@ -23,18 +24,26 @@ async def get_category_products(category_id: int, page: int=0, limit: int = 50):
 
 
 @router.post("/")
-async def create_category(data: CategoryCreateModel):
+async def create_category(data: CategoryCreateModel, manager: AuthManagerDep):
+    if not Permissions.check(manager, Permissions.MANAGE_CATEGORIES):
+        raise HTTPException(status_code=403, detail="Insufficient privileges!")
+
     return await Category.create(name=data.name, description=data.description)
 
 
 @router.patch("/{category_id}")
-async def update_category(category_id: int, data: CategoryUpdateModel):
+async def update_category(category_id: int, data: CategoryUpdateModel, manager: AuthManagerDep):
+    if not Permissions.check(manager, Permissions.MANAGE_CATEGORIES):
+        raise HTTPException(status_code=403, detail="Insufficient privileges!")
+
     cat = await Category.get_or_none(id=category_id)
     await cat.update(**data.model_dump(exclude_defaults=True))
-
     return cat
 
 
 @router.delete("/{category_id}", status_code=204)
-async def delete_category(category_id: int):
+async def delete_category(category_id: int, manager: AuthManagerDep):
+    if not Permissions.check(manager, Permissions.MANAGE_CATEGORIES):
+        raise HTTPException(status_code=403, detail="Insufficient privileges!")
+
     await Category.filter(id=category_id).delete()
